# Speed up time encoding/decoding