###### Required for all non-trivial PRs
- [ ] Rebased/mergable
- [ ] Tests pass
- [ ] CHANGELOG.md updated
- [ ] Sign [CLA](https://influxdata.com/community/cla/) (if not already signed)

This speeds up time encoding and decoding by skipping the divisor
scaling if scaling by 1.  Since division and multiplication are expensive
cpu and scaling by 1 has no effect, this just slows encoding and decoding
down.

This should improve compaction times somewhat since time encoding is used by all types.

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkTimeEncoder-8            33930         19271         -43.20%
BenchmarkTimeDecoder_Packed-8     9819          8339          -15.07%
BenchmarkTimeDecoder_RLE-8        2331          2288          -1.84%
```